dashboard/requirements.txt
altair
datetime
ipykernel
pandas
pyarrow
pylint
python-dotenv
redshift-connector
seaborn
streamlit